Added destructible plates and raycasting collision for bullets
The plates are thin and should be hit by bullets. Because of the fast travel speed of bullets, the updated position may be on the other side. Therefore, a raycast between the last position and the current position is implemented and the object is moved back to where the collision should have occurred to ensure that the collision does indeed occur.