Plates now follow node-based path
A node array is built upon run time which is passed to the plates that are spawned. They use a combination of force and torque pushes to orient themselves along the path defined by the nodes given. The plates now also only get destroyed after a number of hits (losing opacity each time) and only get destroyed on hits from bullets (not other plates).